# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mass Polimer ERP API",
    "version": "1.0.0",
...
    "license": {
      "name": "UNLICENSED"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "security": [
    {
      "FirebaseAuth": []
    },
    {
      "DevUser": []
    }
  ],
  "paths": {
...
    "/api/formulations": {
      "get": {
        "tags": [
          "Formulations"
        ],
        "operationId": "listFormulations",
        "summary": "List formulations",
        "description": "**Requires permission:** `screen:formulations`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Formulations with their components.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Formulation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:formulations`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Formulations"
        ],
        "operationId": "createFormulation",
        "summary": "Create a formulation",
        "description": "**Requires permission:** `action:formula.edit`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "minLength": 1
                  },
                  "product": {
                    "type": "string",
                    "default": ""
                  },
                  "components": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "pct": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 100
                        },
                        "lotId": {
                          "type": "string",
                          "default": ""
                        }
                      },
                      "required": [
                        "name",
                        "pct"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "code",
                  "components"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created formulation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Formulation"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `action:formula.edit`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "`validation` — the body failed schema validation. `error.details` carries the field errors.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/formulations/{id}": {
      "patch": {
        "tags": [
          "Formulations"
        ],
        "operationId": "updateFormulation",
        "summary": "Update a formulation",
        "description": "**Requires permission:** `action:formula.edit`",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Formulation id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product": {
                    "type": "string"
                  },
                  "components": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "pct": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 100
                        },
                        "lotId": {
                          "type": "string",
                          "default": ""
                        }
                      },
                      "required": [
                        "name",
                        "pct"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1
                  },
                  "active": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated formulation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Formulation"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `action:formula.edit`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "- `not_found` — No formulation with that id exists in your organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "- `locked` — The revision is locked and cannot be edited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "`validation` — the body failed schema validation. `error.details` carries the field errors.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "FirebaseAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token, verified server-side. This is the production scheme; it activates when `DEV_AUTH=0` and the Firebase Admin credentials are configured."
      },
      "DevUser": {
        "type": "apiKey",
        "in": "header",
        "name": "x-dev-user",
        "description": "Development only, active while `DEV_AUTH` is on. Pass an employee code or email to act as that member; omit it to fall back to a seeded Administrator. Pair with `x-dev-org` to choose the organization. Never enable this in production."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The envelope every failure is returned in.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable code. Branch on this, not on the message."
              },
              "message": {
                "type": "string",
                "description": "Human-readable explanation, safe to surface in a UI."
              },
              "details": {
                "description": "Optional context. For `validation` this is the Zod field-error map."
              }
            }
          }
        }
      },
...
      "Formulation": {
        "type": "object",
        "description": "The Formulation record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "rev": {
            "type": "integer"
          },
          "product": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "locked": {
            "type": "boolean"
          },
          "lockReason": {
            "type": "string"
          },
          "capaId": {
            "type": [
              "string",
              "null"
            ]
          },
          "components": {
            "description": "Arbitrary JSON."
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "code",
          "rev",
          "product",
          "active",
          "locked",
          "lockReason",
          "components",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}